+++ source/MultiFactorialStatsUpdater.cpp
@@ -67,7 +67,7 @@ bool  MultiFactorialStatsUpdater::UpdateGroupStats(string candidateName)
    return UpdateGroupStats(0, theB, 0, theC);
 }
 
-bool  MultiFactorialStatsUpdater::UpdateGroupStats(int64_t theK, int32_t theB, int32_t theN, int32_t theC)
+bool  MultiFactorialStatsUpdater::UpdateGroupStats(int64_t theK, int32_t theB, int32_t theN, int64_t theC)
 {
    SQLStatement *sqlStatement;
    bool          success;
@@ -149,9 +149,9 @@ bool  MultiFactorialStatsUpdater::UpdateGroupStats(int64_t theK, int32_t theB, i
    // tested.  Note that the $null_func$ is needed in case only one candidate in the group
    // has been tested.  In that case it returns that candidate.
    if (nextToTest == 0)
-      snprintf(completedSQL, sizeof(completedSQL), "(select max(n) from Candidate where b = %d and c = %d)", theB, theC);
+      snprintf(completedSQL, sizeof(completedSQL), "(select max(n) from Candidate where b = %d and c = %" PRId64")", theB, theC);
    else
-      snprintf(completedSQL, sizeof(completedSQL), "$null_func$((select max(n) from Candidate where b = %d and c = %d and n < %d), %d)",
+      snprintf(completedSQL, sizeof(completedSQL), "$null_func$((select max(n) from Candidate where b = %d and c = %" PRId64" and n < %d), %d)",
               theB, theC, nextToTest, nextToTest);
 
    // Finally, update the group stats
@@ -168,7 +168,7 @@ bool  MultiFactorialStatsUpdater::UpdateGroupStats(int64_t theK, int32_t theB, i
 }
 
 bool   MultiFactorialStatsUpdater::InsertCandidate(string candidateName, int64_t theK, int32_t theB, int32_t theN,
-                                                   int32_t theC, int32_t theD, double decimalLength)
+                                                   int64_t theC, int32_t theD, double decimalLength)
 {
    const char *insertSQL = "insert into Candidate " \
                            "( CandidateName, DecimalLength, b, n, c, LastUpdateTime ) " \
